+++ Forge.Forms/src/Forge.Forms.Demo/Routes/ExamplesRoute.cs
@@ -118,6 +118,8 @@ namespace Forge.Forms.Demo.Routes
 
             yield return new ExamplePresenter(new TextFields(), "Text Fields", large);
 
+            yield return new ExamplePresenter(new TextElements(), "Text Elements", large);
+
             yield return new ExamplePresenter(new Selection(), "Selection", large);
 
             yield return new ExamplePresenter(new FoodSelection(), "Food Selection", large);